Moved vernacular query to page.js, simplified page.svelte
